requirements.txt
PyQt6>=6.6.1
mutagen>=1.47.0
Pillow>=10.2.0
